Got it working with Regular 8, 48bit HDR, using some OpenCV.

+++ s8/src/main/java/ImageDisplay.java
@@ -86,7 +86,7 @@ public class ImageDisplay {
 			public void mouseDown(MouseEvent e) {
 		        Object pixelArray = raster.getDataElements(e.x, e.y, (Object)null);
 
-            	System.out.println("(" + e.x + ", " + e.y + ")" + Arrays.toString((short[]) pixelArray));
+            	System.out.println("(" + e.x + ", " + e.y + ")" + Arrays.toString((byte[]) pixelArray));
 			}
 			
 			@Override
